Structuring the code
overall structure done, mathematics need some care....
